Added get post by ID

+++ vue_starter_java/backend/src/main/java/com/techelevator/model/PostDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface PostDao {
 	
- List<Post> getAllPostsByPostId(Long postId);
+ List<Post> getAllPostsByPostId(int postId);
  List<Post> getAllPosts();
 	 
 	 void save(Post savePost);
